// types/tracking/meta/ClientUserData.ts
// Path: types/tracking/meta/ClientUserData.ts
export type ClientUserData = {
  fbp?: string | undefined
  fbc?: string | undefined
  external_id?: string | undefined
  email?: string | undefined
  email_hash?: string | undefined
  phone?: string | undefined
  first_name?: string | undefined
  last_name?: string | undefined
  date_of_birth?: string | undefined
  gender?: string | undefined
  city?: string | undefined
  state?: string | undefined
  zip?: string | undefined
  country?: string | undefined
  client_ip_address?: string | undefined
  client_user_agent?: string | undefined
}